requirements.txt
fastapi==0.111.0
numpy==2.0.0
Pillow==10.4.0
Requests==2.32.3
torch==2.2.2
transformers==4.42.3
